# go/eugene-go-starter/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "登录鉴权最小可用文档",
        "title": "Eugene API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/eugene-go-starter_internal_handler_dto.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.LoginOK"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrResp"
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "刷新当前 token",
                "parameters": [
                    {
                        "description": "刷新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.RefreshDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.LoginOK"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "eugene-go-starter_internal_handler_dto.LoginReq": {
            "type": "object",
            "required": [
                "password",
                "siteId",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "siteId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_handler.ErrResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 10008
                },
                "msg": {
                    "type": "string",
                    "example": "wrong token"
                }
            }
        },
        "internal_handler.LoginOK": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "integer",
                    "example": 1734567890
                },
                "jti": {
                    "description": "可选",
                    "type": "string",
                    "example": "7a1f8d9c4e..."
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                }
            }
        },
        "internal_handler.RefreshDTO": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}